Add style to content
