+++ assets/style.css
@@ -6,6 +6,8 @@ body {
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
+  color: #fff;
+  overflow-y: hidden;
 }
 
 .content-div {
@@ -19,11 +21,24 @@ body {
   margin-top: 20px;
 }
 
-.mainDiv {
+.main-container {
   display: flex;
-  flex-direction: column;
   justify-content: center;
   align-items: center;
+  margin-top: 30px;
+}
+
+p {
+  font-size: 1.2rem;
+}
+
+.search-div {
+  justify-content: center;
+  align-items: center;
+}
+
+.main-container div {
+  color: #fff;
 }
 
 .app-title {
